docker/marker-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
marker-pdf==0.2.1
pydantic==2.5.0
python-multipart==0.0.6